+++ packages/webapp/src/app/sessions/[workerId]/actions.ts
@@ -2,10 +2,11 @@
 
 import { fetchTodoListSchema, sendMessageToAgentSchema, updateAgentStatusSchema } from './schemas';
 import { authActionClient } from '@/lib/safe-action';
-import { PutCommand, UpdateCommand } from '@aws-sdk/lib-dynamodb';
+import { PutCommand } from '@aws-sdk/lib-dynamodb';
 import { ddb, TableName } from '@remote-swe-agents/agent-core/aws';
 import { getOrCreateWorkerInstance, renderUserMessage, getTodoList } from '@remote-swe-agents/agent-core/lib';
-import { MessageItem, sendWorkerEvent, updateSessionAgentStatus } from '@remote-swe-agents/agent-core/lib';
+import { sendWorkerEvent, updateSessionAgentStatus } from '@remote-swe-agents/agent-core/lib';
+import { MessageItem } from '@remote-swe-agents/agent-core/schema';
 
 export const sendMessageToAgent = authActionClient
   .schema(sendMessageToAgentSchema)
